refactor: move sidebar nav out of header and into ExploreIframeViews.js
