+++ pages/explore/TreePath.js
@@ -1,11 +1,7 @@
 import React from 'react';
 import Link from 'next/link';
-import { getCoreRowModel, useReactTable } from '@tanstack/react-table';
-import { RiBrush2Line, RiBrushLine, RiChat1Line, RiEditLine, RiFolderLine, RiFolderUnknowLine, RiLink } from '@remixicon/react';
-import { useRouter } from 'next/router';
 
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@/components/UI/shadcn/DropdownMenu';
-import TreeLeaves from './TreeLeaves';
 import {
     Breadcrumb,
     BreadcrumbEllipsis,
@@ -15,87 +11,8 @@ import {
     BreadcrumbSeparator,
 } from '@/components/UI/shadcn/Breadcrumb';
 import LoadingSpinnerInline from '@/components/UI/LoadingSpinnerInline';
-import Icon from '@/components/UI/Icon';
-import { Table, TableBody } from '@/components/UI/shadcn/Table';
 
-const TreePath = ({ selectedSpaceChildren, isFetchingContent, iframeRoomId }) => {
-    const router = useRouter();
-    const columns = [
-        {
-            accessorKey: 'icon',
-            header: (
-                <Icon>
-                    <RiFolderUnknowLine />
-                </Icon>
-            ),
-            cell: ({ row }) => {
-                if (row.original?.meta?.template === 'etherpad') {
-                    return (
-                        <Icon>
-                            <RiEditLine />
-                        </Icon>
-                    );
-                }
-
-                if (row.original?.meta?.template === 'spacedeck') {
-                    return (
-                        <Icon>
-                            <RiBrush2Line />
-                        </Icon>
-                    );
-                }
-
-                if (row.original?.meta?.template === 'tldraw') {
-                    return (
-                        <Icon>
-                            <RiBrushLine />
-                        </Icon>
-                    );
-                }
-
-                if (row.original?.meta?.template === 'link') {
-                    return (
-                        <Icon>
-                            <RiLink />
-                        </Icon>
-                    );
-                }
-
-                if (row.original?.meta?.type === 'context') {
-                    return (
-                        <Icon>
-                            <RiFolderLine />
-                        </Icon>
-                    );
-                }
-
-                if (!row.original?.meta) {
-                    return (
-                        <Icon>
-                            <RiChat1Line />
-                        </Icon>
-                    );
-                }
-            },
-        },
-        {
-            accessorKey: 'name',
-            header: 'Name',
-            cell: ({ row }) => (
-                <Link target={row.target} href={row.href} rel="noopener noreferrer" className="flex items-center justify-between">
-                    {row.getValue('name')}
-                </Link>
-            ),
-        },
-    ];
-    const data = selectedSpaceChildren[selectedSpaceChildren.length - 1];
-
-    const table = useReactTable({
-        data,
-        columns,
-        getCoreRowModel: getCoreRowModel(),
-    });
-
+const TreePath = ({ selectedSpaceChildren, isFetchingContent }) => {
     return (
         <>
             <Breadcrumb>
@@ -175,22 +92,6 @@ const TreePath = ({ selectedSpaceChildren, isFetchingContent, iframeRoomId }) =>
                     })}
                 </BreadcrumbList>
             </Breadcrumb>
-            {iframeRoomId && table.getRowModel().rows?.length
-                ? table.getRowModel().rows.map((row, index) => {
-                      <Table>
-                          <TableBody>
-                              const roomId = row.original.room_id || row.original.roomId || row.original.id; if (index === 0) return null;
-                              return (
-                              <TreeLeaves
-                                  key={row.id}
-                                  row={row}
-                                  selected={router.query.roomId[1] === roomId || router.query.roomId[0] === roomId}
-                              />
-                              );
-                          </TableBody>
-                      </Table>;
-                  })
-                : null}
         </>
     );
 };
